fix(dash3d): world3d Occluded MinDeltaZ, DrawTile ModelB \!wallVisible, CollisionMap AddWall typos, projectile X grouping
- world3d Occluded mode 5: var5.MinDeltaZ for var9 (Java World3D.java:2285),
  not MinDeltaX; vertical-occluder Z-projection was using X-slope.
- world3d DrawTile front-pass: \!WallVisible for ModelB (Java
  World3D.java:1407); the dropped \! drew ModelB only when fully occluded
  and hid it when visible. The TypeA test on the prior line was correct.
- collisionmap AddWall arg5=0/arg1=3: AddCMap(var8, var7-1, 0x400) (Java
  CollisionMap.java:125); the first arg was var7, putting the east-edge
  block bit on the wrong cell.
- collisionmap AddWall arg5=2/arg1=0: AddCMap(var8, var7, 66560) (Java
  CollisionMap.java:150); the constant was 65560. DelWall mirror was
  already correct.
- projectile UpdateVelocity X: SrcX + var6*Arc/var10 (Java
  ProjectileEntity.java:104); the Go grouping (SrcX+var6)*Arc/var10
  produced an asymmetric X path vs the structurally-correct Z line.

+++ pkg/jagex2/dash3d/entity/projectileentity.go
@@ -57,7 +57,7 @@ func (e *ProjectileEntity) UpdateVelocity(arg0, arg1, arg2, arg4 int) {
 		var6 := arg2 - e.SrcX
 		var8 := float64(arg1 - e.SrcZ)
 		var10 := math.Sqrt(float64(var6*var6) + var8*var8)
-		e.X = float64(e.SrcX+var6) * float64(e.Arc) / var10
+		e.X = float64(e.SrcX) + float64(var6)*float64(e.Arc)/var10
 		e.Z = float64(e.SrcZ) + var8*float64(e.Arc)/var10
 		e.Y = float64(e.SrcY)
 	}
